+++ client/src/app/site/pages/meetings/modules/poll/components/poll-progress/poll-progress.component.ts
@@ -7,8 +7,8 @@ import { ModelRequestBuilderService } from 'src/app/site/services/model-request-
 import { OperatorService } from 'src/app/site/services/operator.service';
 import { UserControllerService } from 'src/app/site/services/user-controller.service';
 import { BaseUiComponent } from 'src/app/ui/base/base-ui-component';
+import { getParticipantVoteInfoSubscriptionConfig } from '../../../../pages/participants/participants.subscription';
 
-import { getParticipantVoteInfoSubscriptionConfig } from '../../../../pages/participants/config/model-subscription';
 import { ActiveMeetingService } from '../../../../services/active-meeting.service';
 
 @Component({
